requirements.txt
# Env
python-dotenv==1.0.1

# Data
polars==0.20.31
pandas==2.2.2
numpy==1.26.4
duckdb==0.10.3

# Ingestion
httpx==0.27.0
supabase==2.5.1
soccerdata==0.5.0
statsbombpy==1.1.3
understat==0.1.1

# Orchestration
prefect==2.19.7

# ML
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.5
scikit-learn==1.5.0
optuna==3.6.1
shap==0.45.1
umap-learn==0.5.6
joblib==1.4.2
pulp==2.8.0

# Viz
mplsoccer==1.3.0
matplotlib==3.9.0
plotly==5.22.0
altair==5.3.0
kaleido==0.2.1

# Dashboard
streamlit==1.35.0

# Notebooks
marimo==0.6.19
